posbys/requirements.txt
django >=1.9, <2
djangorestframework == 3.6.3
SQLAlchemy == 1.1.10
psycopg2 == 2.7.1
django-oauth-toolkit == 0.12.0
django-debug-toolbar == 1.8
django-cors-headers == 2.1.0
requests == 2.18.1